+++ src/pydiverse/pipedag/backend/table/sql.py
@@ -353,6 +353,17 @@ class SQLTableStore(BaseTableStore):
 
         return self._execute(query, conn)
 
+    @engine_dispatch
+    def name_adj(self, name: str):
+        # some dialects need to replace all lowercase names by uppercase because
+        # they create uppercase table names by default
+        return name
+
+    @name_adj.dialect("ibm_db_sa")
+    def _name_adj_ibm_db_sa(self, name: str):
+        # DB2 creates tables uppercase if all lowercase given
+        return name.upper() if name.islower() else name
+
     def add_indexes(
         self, table: Table, schema: Schema, *, early_not_null_possible: bool = False
     ):
@@ -476,7 +487,9 @@ class SQLTableStore(BaseTableStore):
         self, src_table: str, src_schema: Schema, dest_table: str, dest_schema: Schema
     ):
         inspector = sa.inspect(self.engine)
-        pk_constraint = inspector.get_pk_constraint(src_table, schema=src_schema.get())
+        pk_constraint = inspector.get_pk_constraint(
+            self.name_adj(src_table), schema=self.name_adj(src_schema.get())
+        )
         if len(pk_constraint["constrained_columns"]) > 0:
             self.add_primary_key(
                 dest_table,
@@ -484,7 +497,9 @@ class SQLTableStore(BaseTableStore):
                 pk_constraint["constrained_columns"],
                 name=pk_constraint["name"],
             )
-        indexes = inspector.get_indexes(src_table, schema=src_schema.get())
+        indexes = inspector.get_indexes(
+            self.name_adj(src_table), schema=self.name_adj(src_schema.get())
+        )
         for index in indexes:
             if len(index["include_columns"]) > 0:
                 self.logger.warning(
@@ -511,7 +526,9 @@ class SQLTableStore(BaseTableStore):
         self, col_names: Iterable[str], table_name: str, schema: Schema
     ):
         inspector = sa.inspect(self.engine)
-        columns = inspector.get_columns(table_name, schema=schema.get())
+        columns = inspector.get_columns(
+            self.name_adj(table_name), schema=self.name_adj(schema.get())
+        )
         types = {d["name"]: d["type"] for d in columns}
         sql_types = [types[col] for col in col_names]
         return sql_types
@@ -652,9 +669,13 @@ class SQLTableStore(BaseTableStore):
                 # Attention: similar code in sql_ddl.py:visit_drop_schema_ibm_db_sa
                 # for drop.cascade=True
                 inspector = sa.inspect(self.engine)
-                for table in inspector.get_table_names(schema=transaction_schema.get()):
+                for table in inspector.get_table_names(
+                    schema=self.name_adj(transaction_schema.get())
+                ):
                     self.execute(DropTable(table, schema=transaction_schema))
-                for view in inspector.get_view_names(schema=transaction_schema.get()):
+                for view in inspector.get_view_names(
+                    schema=self.name_adj(transaction_schema.get())
+                ):
                     self.execute(DropView(view, schema=transaction_schema))
                 self.drop_all_dialect_specific(schema=transaction_schema)
             else:
@@ -865,7 +886,9 @@ class SQLTableStore(BaseTableStore):
         _ = src_schema  # only needed by other dialects
         for table in src_tables:
             self.execute(
-                CreateViewAsSelect(table.name, dest_schema, sa.select(table)),
+                CreateViewAsSelect(
+                    table.name, dest_schema, sa.select("*").select_from(table)
+                ),
                 conn=conn,
             )
 
@@ -902,7 +925,9 @@ class SQLTableStore(BaseTableStore):
     def copy_table_to_transaction(self, table: Table):
         stage = table.stage
         schema_name = self.get_schema(stage.name).get()
-        has_table = sa.inspect(self.engine).has_table(table.name, schema=schema_name)
+        has_table = sa.inspect(self.engine).has_table(
+            self.name_adj(table.name), schema=self.name_adj(schema_name)
+        )
         if not has_table:
             raise RuntimeError(
                 f"Can't copy table '{table.name}' (schema: '{stage.name}')"
@@ -1009,7 +1034,9 @@ class SQLTableStore(BaseTableStore):
     def copy_lazy_table_to_transaction(self, metadata: LazyTableMetadata, table: Table):
         stage = table.stage
         schema_name = self.get_schema(metadata.stage).get()
-        has_table = sa.inspect(self.engine).has_table(metadata.name, schema=schema_name)
+        has_table = sa.inspect(self.engine).has_table(
+            self.name_adj(metadata.name), schema=self.name_adj(schema_name)
+        )
         if not has_table:
             msg = (
                 f"Can't copy lazy table '{metadata.name}' (schema:"
@@ -1063,7 +1090,7 @@ class SQLTableStore(BaseTableStore):
         """
         _ = include_everything  # not used in this implementation
         inspector = sa.inspect(self.engine)
-        return inspector.get_view_names(schema)
+        return inspector.get_view_names(self.name_adj(schema))
 
     @get_view_names.dialect("mssql")
     def _get_view_names_mssql(self, schema: str, *, include_everything=False):
@@ -1405,7 +1432,7 @@ class SQLTableStore(BaseTableStore):
         inspector = sa.inspect(self.engine)
         schema = self.get_schema(stage.transaction_name).get()
 
-        table_names = inspector.get_table_names(schema)
+        table_names = inspector.get_table_names(self.name_adj(schema))
         view_names = self.get_view_names(schema, include_everything=include_everything)
 
         return table_names + view_names
@@ -1451,8 +1478,8 @@ class SQLAlchemyTableHook(TableHook[SQLTableStore]):
         task_info: TaskInfo,
     ):
         obj = table.obj
-        if isinstance(table.obj, sa.Table):
-            obj = sa.select(table.obj)
+        if isinstance(table.obj, (sa.Table, sa.sql.selectable.Alias)):
+            obj = sa.select("*").select_from(table.obj)
 
         source_tables = [
             dict(
@@ -1478,7 +1505,7 @@ class SQLAlchemyTableHook(TableHook[SQLTableStore]):
         store.add_indexes(table, schema, early_not_null_possible=True)
 
     @classmethod
-    def retrieve(cls, store, table, stage_name, as_type):
+    def retrieve(cls, store, table, stage_name, as_type, namer):
         table_name = table.name
         schema = store.get_schema(stage_name).get()
         table_name, schema = store.resolve_aliases(table_name, schema)
@@ -1490,7 +1517,7 @@ class SQLAlchemyTableHook(TableHook[SQLTableStore]):
                     sa.MetaData(),
                     schema=schema,
                     autoload_with=store.engine,
-                )
+                ).alias(namer.get_name(table_name))
                 break
             except (sa.exc.SQLAlchemyError, sa.exc.DBAPIError):
                 if retry_iteration == 3:
@@ -1501,7 +1528,7 @@ class SQLAlchemyTableHook(TableHook[SQLTableStore]):
     @classmethod
     def lazy_query_str(cls, store, obj) -> str:
         if isinstance(obj, sa.sql.selectable.FromClause):
-            query = sa.select(sa.text("*")).select_from(obj)
+            query = sa.select("*").select_from(obj)
         else:
             query = obj
         query_str = str(
@@ -1725,7 +1752,7 @@ class PandasTableHook(TableHook[SQLTableStore]):
         return cols, year_cols
 
     @classmethod
-    def retrieve(cls, store, table, stage_name, as_type):
+    def retrieve(cls, store, table, stage_name, as_type, namer):
         schema = store.get_schema(stage_name).get()
         with store.engine.connect() as conn:
             table_name = table.name
@@ -1823,7 +1850,7 @@ class PolarsTableHook(TableHook[SQLTableStore]):
         store.add_indexes(table, schema)
 
     @classmethod
-    def retrieve(cls, store, table, stage_name, as_type):
+    def retrieve(cls, store, table, stage_name, as_type, namer):
         schema = store.get_schema(stage_name).get()
         table_name = table.name
         table_name, schema = store.resolve_aliases(table_name, schema)
@@ -1867,8 +1894,8 @@ class TidyPolarsTableHook(TableHook[SQLTableStore]):
         PolarsTableHook.materialize(store, table, stage_name, task_info)
 
     @classmethod
-    def retrieve(cls, store, table, stage_name, as_type):
-        df = PolarsTableHook.retrieve(store, table, stage_name, as_type)
+    def retrieve(cls, store, table, stage_name, as_type, namer):
+        df = PolarsTableHook.retrieve(store, table, stage_name, as_type, namer)
         return tidypolars.from_polars(df)
 
     @classmethod
@@ -1919,15 +1946,17 @@ class PydiverseTransformTableHook(TableHook[SQLTableStore]):
         raise NotImplementedError
 
     @classmethod
-    def retrieve(cls, store, table, stage_name, as_type):
+    def retrieve(cls, store, table, stage_name, as_type, namer):
         from pydiverse.transform.eager import PandasTableImpl
         from pydiverse.transform.lazy import SQLTableImpl
 
         if issubclass(as_type, PandasTableImpl):
-            df = PandasTableHook.retrieve(store, table, stage_name, pd.DataFrame)
+            df = PandasTableHook.retrieve(store, table, stage_name, pd.DataFrame, namer)
             return pdt.Table(PandasTableImpl(table.name, df))
         if issubclass(as_type, SQLTableImpl):
-            sa_tbl = SQLAlchemyTableHook.retrieve(store, table, stage_name, sa.Table)
+            sa_tbl = SQLAlchemyTableHook.retrieve(
+                store, table, stage_name, sa.Table, namer
+            )
             return pdt.Table(SQLTableImpl(store.engine, sa_tbl))
         raise NotImplementedError
 
@@ -2007,7 +2036,7 @@ class IbisTableHook(TableHook[SQLTableStore]):
         return SQLAlchemyTableHook.materialize(store, table, stage_name, task_info)
 
     @classmethod
-    def retrieve(cls, store, table, stage_name, as_type):
+    def retrieve(cls, store, table, stage_name, as_type, namer):
         con = cls.get_con(store)
         table_name = table.name
         schema = store.get_schema(stage_name).get()