+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ https://semver.org/spec/v2.0.0.html. The most important bits are:
 Changelog
 =========
 
+0.2.5 (YYYY-MM-DD)
+------------------
+- fix ibm_db_sa bug when copying dataframes from cache: uppercase table names by default
+- nicer readable SQL queries: use automatic aliases for inputs of SQLAlchemy tasks
+
 0.2.4 (2023-05-05)
 ------------------
 - fix errors by increasing output_json length in metadata table